fix(wick-wasm): align Cargo.toml include comment + simplify length type in chat-message parse
Two PR-review nits from PR #72 review pass on 3dd6d11:

1. wick-wasm/Cargo.toml: the comment block above `include = [...]`
   still said README.md "is generated as `pkg/README.md` ... so we
   don't list it here" — but the prior commit added README.md to
   the include list. Rewrote the comment to describe what the list
   actually does: README is listed for crates.io publication, and
   wasm-pack also copies it to pkg/README.md for the npm side.
   Symlink note clarified to mention cargo + npm + git transparency.

2. parse_chat_messages: `let len = array.length() as usize` then
   looping `for i in 0..len as u32` did a needless cast roundtrip.
   `js_sys::Array::length` returns u32 and `Array::get` takes u32 —
   keep len in u32 for the loop and only widen to usize at the
   Vec::with_capacity call. Comment notes the type alignment.

Native build, fmt, clippy --workspace -D warnings clean.

+++ wick-wasm/src/lib.rs
@@ -345,11 +345,14 @@ fn parse_chat_messages(value: &JsValue) -> Result<Vec<wick::tokenizer::ChatMessa
     let array = value
         .dyn_ref::<js_sys::Array>()
         .ok_or_else(|| JsError::new("messages must be an array"))?;
-    let len = array.length() as usize;
-    let mut msgs = Vec::with_capacity(len);
+    // `js_sys::Array::length` returns `u32` and that's the index
+    // type `Array::get` takes — keep `len` in `u32` for the loop
+    // and only widen to `usize` at the `Vec::with_capacity` call.
+    let len = array.length();
+    let mut msgs = Vec::with_capacity(len as usize);
     let role_key = JsValue::from_str("role");
     let content_key = JsValue::from_str("content");
-    for i in 0..len as u32 {
+    for i in 0..len {
         let entry = array.get(i);
         let role = js_sys::Reflect::get(&entry, &role_key)
             .map_err(|_| JsError::new(&format!("messages[{i}] missing role")))?
